requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
insightface==0.7.3
mysql-connector-python==8.0.26
numpy==1.21.2
python-multipart==0.0.5
PyYAML==6.0
scikit-learn==1.0
onnxruntime==1.12.0